Add files via upload
Added a smoother version of the previously implemented code which used normals to determine which faces of the cube were facing the camera, and hence which should be displayed. This version is smoother and is possibly a final solution. One thing which should be mentioned as a note to myself is that the camera used in this version is at the coordinates (0, 0, 0). Idk if this is important yet, but could be to myself later