Added object files.
Whilst the object files do work and can be drawn individually by adjusting the code within the "on_user_update()" function, there are issues with the colouring / shapes themselves in that there are negative dot_products being fed into the colour_scale() function. The reason for this needs investigating and patching, and so far has only been seen with the torus object, and on previous testing with a monkey object. It may be due to the depth of these shapes(?). Needs more looking into.